ENH: for easier testing, avoid re-download of prostate mrb
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=21364
git-svn-id: http://svn.slicer.org/Slicer4/trunk@21364 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/SlicerApp/Testing/Python/RSNA2012ProstateDemo.py
@@ -188,8 +188,11 @@ class RSNA2012ProstateDemoTest(unittest.TestCase):
     # perform the downloads if needed, then load
     filePath = slicer.app.temporaryPath + '/RSNA2012ProstateDemo.mrb'
     urlPath = 'http://slicer.kitware.com/midas3/download?items=10697'
-    self.delayDisplay('Downloading MRB from %s to %s...\n' % (urlPath, filePath))
-    urllib.urlretrieve(urlPath, filePath)
+    if not os.path.exists(filePath) or os.stat(filePath).st_size == 0:
+      self.delayDisplay('Downloading MRB from %s to %s...\n' % (urlPath, filePath))
+      urllib.urlretrieve(urlPath, filePath)
+    else:
+      self.delayDisplay('Using existing %s...\n' % filePath)
     slicer.mrmlScene.Clear(0)
     appLogic = slicer.app.applicationLogic()
     self.delayDisplay('Done loading data! Will now open the bundle')
